Create Server.cpp
This is a modified version of Lab 2 that echoes to all connected clients.  The hitch with this iteration is that it waits for certain clients to send before echoing the message to everyone.  i.e. Everyone may need to enter a message to get a message back.  When I have a bit more time I'll compose this into proper functions and a class.

The Lab 3 version of this, with threads, lacks this hitch.  Does anyone know if Miller forbade the use of threads on this?